fix(ai-rerank): address rerank provider review issues
