+++ packages/ai-sdk-provider/src/index.ts
@@ -1,1 +1,2 @@
 export * from './cherryin-provider'
+export * from './openai-compatible-reranking-model'
