+++ packages/aiCore/src/core/runtime/__tests__/types.test.ts
@@ -1,7 +1,8 @@
 import type { JSONObject } from '@ai-sdk/provider'
 import { describe, expectTypeOf, it } from 'vitest'
 
-import type { RerankResult } from '../types'
+import type { RuntimeExecutor } from '../executor'
+import type { RerankParams, RerankResult } from '../types'
 
 describe('Runtime rerank types', () => {
   it('defaults rerank documents to string', () => {
@@ -15,5 +16,13 @@ describe('Runtime rerank types', () => {
     type Document = RerankResult<DocumentValue>['ranking'][number]['document']
 
     expectTypeOf<Document>().toEqualTypeOf<DocumentValue>()
+  })
+
+  it('preserves object document generics on RuntimeExecutor.rerank', () => {
+    type DocumentValue = JSONObject & { text: string }
+
+    expectTypeOf<RuntimeExecutor['rerank']>().toMatchTypeOf<
+      (params: RerankParams<DocumentValue>) => Promise<RerankResult<DocumentValue>>
+    >()
   })
 })
