+++ src/main/ai/AiService.ts
@@ -133,7 +133,6 @@ export interface AiRerankResult {
   ranking: Array<{
     originalIndex: number
     score: number
-    document: string
   }>
 }
 
@@ -530,8 +529,7 @@ export class AiService extends BaseService {
     return {
       ranking: result.ranking.map((item) => ({
         originalIndex: item.originalIndex,
-        score: item.score,
-        document: request.documents[item.originalIndex]
+        score: item.score
       }))
     }
   }
@@ -576,7 +574,12 @@ export class AiService extends BaseService {
     }
     let probe: Promise<unknown>
     if (isRerankModel(model)) {
-      probe = this.rerank({ ...probeRequest, query: 'test', documents: ['test'], topN: 1 })
+      probe = this.rerank({ ...probeRequest, query: 'test', documents: ['test'], topN: 1 }).then((result) => {
+        if (result.ranking.length === 0) {
+          throw new Error('Rerank health check returned empty ranking')
+        }
+        return result
+      })
     } else if (isEmbeddingModel(model)) {
       probe = this.embedMany({ ...probeRequest, values: ['test'] })
     } else {
